fix: SCRUM-147 삭제 다이얼로그 중앙에 뜨게 수정
