+++ presentation/src/main/java/com/pomonyang/mohanyang/presentation/screen/home/category/component/CategoryBottomSheetContents.kt
@@ -7,12 +7,6 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateMapOf
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
@@ -25,68 +19,30 @@ import com.pomonyang.mohanyang.presentation.designsystem.button.box.MnBoxButton
 import com.pomonyang.mohanyang.presentation.designsystem.button.box.MnBoxButtonColorType
 import com.pomonyang.mohanyang.presentation.designsystem.button.box.MnBoxButtonStyles
 import com.pomonyang.mohanyang.presentation.designsystem.button.select.MnSelectListItem
-import com.pomonyang.mohanyang.presentation.designsystem.dialog.MnDialog
 import com.pomonyang.mohanyang.presentation.designsystem.token.MnSpacing
 import com.pomonyang.mohanyang.presentation.model.category.PomodoroCategoryModel
 import com.pomonyang.mohanyang.presentation.screen.home.category.model.CategoryIcon
 import com.pomonyang.mohanyang.presentation.screen.home.category.model.CategoryManageState
 import com.pomonyang.mohanyang.presentation.theme.MnTheme
 import kotlinx.collections.immutable.ImmutableList
+import kotlinx.collections.immutable.ImmutableMap
 import kotlinx.collections.immutable.persistentListOf
+import kotlinx.collections.immutable.persistentMapOf
 
 @Composable
 fun CategoryBottomSheetContents(
     categoryList: ImmutableList<PomodoroCategoryModel>,
+    selectedItems: ImmutableMap<Int, Boolean>,
     categoryManageState: CategoryManageState,
-    currentSelectedCategoryNo: Int,
     onCategorySelected: (PomodoroCategoryModel) -> Unit,
     onCategoryEdit: (PomodoroCategoryModel) -> Unit,
-    onDeleteItemsClick: (List<Int>) -> Unit,
     onSnackbarShow: (String) -> Unit,
+    onDeleteItemClick: () -> Unit,
+    onSelectItem: (Int) -> Unit,
     modifier: Modifier = Modifier,
 ) {
-    val selectedItems: MutableMap<Int, Boolean> = remember(categoryList) { mutableStateMapOf<Int, Boolean>() }
     val gridCells = if (categoryList.size == 1) GridCells.Fixed(1) else GridCells.Fixed(2)
     val context = LocalContext.current
-    var showDeleteAlertDialog by remember { mutableStateOf(false) }
-
-    if (showDeleteAlertDialog) {
-        MnDialog(
-            title = stringResource(R.string.dialog_delete_category_title),
-            subTitle = stringResource(R.string.dialog_delete_category_subtitle),
-            positiveButton = {
-                MnBoxButton(
-                    modifier = Modifier.fillMaxWidth(),
-                    text = stringResource(R.string.common_cancel),
-                    onClick = { showDeleteAlertDialog = false },
-                    colors = MnBoxButtonColorType.tertiary,
-                    styles = MnBoxButtonStyles.medium,
-                )
-            },
-            negativeButton = {
-                MnBoxButton(
-                    modifier = Modifier.fillMaxWidth(),
-                    text = stringResource(R.string.delete_category),
-                    onClick = {
-                        showDeleteAlertDialog = false
-                        onDeleteItemsClick(selectedItems.map { it.key })
-                    },
-                    colors = MnBoxButtonColorType.secondary,
-                    styles = MnBoxButtonStyles.medium,
-                )
-            },
-            onDismissRequest = {
-                showDeleteAlertDialog = false
-            },
-        )
-    }
-
-    LaunchedEffect(categoryManageState) {
-        selectedItems.clear()
-        if (categoryManageState.isDefault()) {
-            selectedItems[currentSelectedCategoryNo] = true
-        }
-    }
 
     Column(modifier = modifier) {
         LazyVerticalGrid(
@@ -120,7 +76,7 @@ fun CategoryBottomSheetContents(
 
                             CategoryManageState.DELETE -> {
                                 if (isNonEditableItem.not()) {
-                                    selectedItems[item.categoryNo] = !selected
+                                    onSelectItem(item.categoryNo)
                                 } else {
                                     onSnackbarShow(context.getString(R.string.default_category_cannot_edit))
                                 }
@@ -139,7 +95,7 @@ fun CategoryBottomSheetContents(
             MnBoxButton(
                 modifier = Modifier.fillMaxWidth(),
                 text = stringResource(id = R.string.change_category_delete_count, selectedItems.values.count { it }),
-                onClick = { showDeleteAlertDialog = true },
+                onClick = onDeleteItemClick,
                 colors = buttonColor,
                 isEnabled = isEnabled,
                 styles = MnBoxButtonStyles.large,
@@ -209,10 +165,11 @@ private fun CategoryBottomSheetPreview(
         CategoryBottomSheetContents(
             categoryList = categoryList,
             categoryManageState = CategoryManageState.EDIT,
-            currentSelectedCategoryNo = 1,
+            selectedItems = persistentMapOf(),
+            onDeleteItemClick = {},
+            onSelectItem = {},
             onCategorySelected = {},
             onCategoryEdit = {},
-            onDeleteItemsClick = {},
             onSnackbarShow = {},
         )
     }
